Make workspaces effectively secret
Use simple scheme of generating an identifier consisting of random words
for each new WorkSpace and using this rather than the ID in URLs.

This should effectively keep a WorkSpace secret and only accessible to
those its URL is shared with, since these identifiers should be random
enough to be un-guessable.

I could be wrong but I think this should also be no less secure than the
previous plan of having a shared password for the WorkSpace, since
either way leaking a single piece of information provides access to the
WorkSpace, while also being easier to implement and easier for
campaigners on the ground to use (a single URL just needs to be followed
rather than a URL and entering a password).

Since an XKCD password-style method of generating this identifier is
used (see https://www.xkcd.com/936) this URL also has an additional
benefit of being fairly easily verbally shareable by campaigners on the
ground, which could turn out to be useful.

For these reasons I think this should be a sufficient method of making
WorkSpaces secret for the MVP, and could also turn out to be sufficient
long term (also relevant for this I believe is that WorkSpaces only need
to remain secret for a day or so before they will be locked down).