Migrate to add WorkSpacePollingStations
Introduce this new model to represent the PollingStation for a
particular WorkSpace (to replace sharing PollingStations between
WorkSpaces, which would be insecure and will not work with incoming
changes).

This commit should introduce the correct model and relationships etc.,
and make all tests pass, but the app is probably still very broken.