Correct WARP count form markup
Previously we had forms (the invalidate forms) inside a form (the new
WARP count form), which doesn't work quite right and leads to weird
behaviour. This commit adapts the markup so the form tag is independent
of the elements in it, which are associated with it via the `form`
attribute (see https://stackoverflow.com/a/21900324/2620402).