Configure SimpleCov
Track and group coverage of all Rake tasks (previously untracked due to
non-standard extension); group coverage for ActiveAdmin configs; track
branch as well as line coverage.

Also needed to upgrade SimpleCov to track the latter, and also split out
a separate `.simplecov` file for this configuration so don't need to
define this in both places we currently start SimpleCov.